Modified Makefile to include headers and specify test
tjha
The Makefile now includes src path headers during compilation. This will
allow for developers to avoid specifying the relative path of header
files and can just specify the name of the header file. The Makefile has
also been changed so that a specific test case can be run directly.