Write utilities for converting to and from UTF-8
jasina, lougheem
This was accidentally written in a different branch
(feature/utf/iterators), so it is all copied here in one big commit
(sorry).
The utilities written allow conversion between pretty much all of our
interally-defined types and a UTF-8-style encoding (stored in a vector
of unsigned chars). This will allow for storing data structures to files
fairly easily.